move functions to the correct places; use utils.to_unicode; cache local timezone offset
